+++ app/src/main/res/layout/activity_movie_screen.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.MovieListActivity">
+    tools:context=".presentation.screen.MovieScreenActivity">
 
     <ListView
         android:id="@+id/movie_list_view"
